购物车列表，购物车全选，全反选，单选，单反选，购物车总数量开发

+++ src/main/java/cn/hust/hustmall/dto/CartProductDTO.java
@@ -37,7 +37,7 @@ public class CartProductDTO {
     @JsonProperty("productStatus")
     private Integer status;
 
-    private Integer productTotalPrice;
+    private BigDecimal productTotalPrice;
 
     @JsonProperty("productStock")
     private Integer stock;
